04_use_cases/legal-doc-assistant/requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
langgraph==0.0.10
python-dotenv==1.0.0
pydantic>=1.7.4,<2.0.0
fastapi==0.99.1
uvicorn==0.22.0
python-multipart==0.0.6
pypdf==3.17.0
chromadb==0.4.0
sentence-transformers==2.2.2
pytest==7.4.0
black==23.7.0
isort==5.12.0
mypy==1.5.0
torch==2.0.1
transformers==4.30.2